+++ src/FlatBuffers.jl
@@ -1,62 +1,5 @@
 module FlatBuffers
 
-import Parameters
-
-function getdef(typedef::Expr)
-    typedef.args[2].args[3]
-end
-
-function getfielddefs(typedef::Expr)
-    getdef(typedef).args[3].args[2:2:(end-1)]
-end
-
-function getinnerconstructor(typedef::Expr)
-    # TODO: this is so evil and will fall over in the slightest breeze.
-    getdef(typedef).args[end].args[end-1].args[end-1]
-end
-
-function getkwvalue(parameters, name)
-    value = nothing
-    for arg in parameters.args
-        @assert arg.head == :kw
-        @assert length(arg.args) == 2
-        if arg.args[1] == name
-            value = arg.args[2]
-        end
-    end
-    return value
-end
-
-function createdefaultfns(typedef::Expr)
-    cons = getinnerconstructor(typedef)
-    parameters = cons.args[end]
-    @assert parameters.head == :parameters
-
-    kwargs = []
-    defs = getfielddefs(typedef)
-    kwdict = Dict{Any, Any}()
-    for d in defs
-        name = d.args[1]
-        type = d.args[end]
-        value = getkwvalue(parameters, name)
-        default = get(kwdict, type, Dict{Symbol, Any}())
-        default[name] = value
-        kwdict[type] = default
-    end
-
-    T = cons.args[1]
-
-    [:(FlatBuffers.default($T, ::Type{$TT}, sym) = eval($(kwdict[TT])[sym]))
-     for TT in keys(kwdict)]
-end
-
-macro with_kw(typedef)
-    body = Parameters.with_kw(typedef, __module__, true)
-    defaults = createdefaultfns(body)
-    defaultsblock = Expr(:block, body, defaults...)
-    return esc(defaultsblock)
-end
-
 # utils
 struct UndefinedType end
 const Undefined = UndefinedType()
